external-client/src/main/java/com/han/oose/rental/data/services/UserServiceImpl.java
package com.han.oose.rental.data.services;

import com.han.oose.rental.data.User;
import com.han.oose.rental.data.repositories.UserRepository;
import com.han.oose.rental.domain.interfaces.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;
import org.springframework.transaction.annotation.Transactional;

import java.util.Optional;

@Component
public class UserServiceImpl implements UserService {

    @Autowired
    private UserRepository userRepository;

    @Override
    public Optional<User> findById(long id) {
        return userRepository.findById(id);
    }

    @Override
    @Transactional
    public void update(User user) {
        userRepository.saveAndFlush(user);
    }

    @Override
    public void setUserRepository(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

}
